lifts the Rendering components into separate module, so they can be shared with other app modules
