+++ app/build.gradle
@@ -41,10 +41,11 @@ dependencies {
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     implementation 'com.google.android.gms:play-services-base:17.5.0'
-    implementation 'androidx.palette:palette-ktx:1.0.0'
 
     implementation "com.google.dagger:hilt-android:$hilt_version"
     kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
+
+    implementation project(path: ':watchfacerenderer')
 
     compileOnly 'com.google.android.wearable:wearable:2.8.1'
 }
