+++ settings.gradle
@@ -1,2 +1,3 @@
+include ':watchfacerenderer'
 include ':app'
 rootProject.name = "Watch Application"
