+++ watchfacerenderer/src/main/res/drawable-nodpi/preview_analog.png
