+++ watchfacerenderer/src/main/res/drawable-nodpi/watchface_service_bg.png
